Working array streaming interface?
Also I guess I don't care about excluding generated files any more.

+++ vivado/ip_repo/export/drivers/byte_count_v1_0/src/xbyte_count.h
@@ -1,6 +1,7 @@
 // ==============================================================
-// Vitis HLS - High-Level Synthesis from C, C++ and OpenCL v2021.2 (64-bit)
-// Copyright 1986-2021 Xilinx, Inc. All Rights Reserved.
+// Vitis HLS - High-Level Synthesis from C, C++ and OpenCL v2022.1 (64-bit)
+// Tool Version Limit: 2022.04
+// Copyright 1986-2022 Xilinx, Inc. All Rights Reserved.
 // ==============================================================
 #ifndef XBYTE_COUNT_H
 #define XBYTE_COUNT_H
@@ -38,7 +39,7 @@ typedef uint64_t u64;
 #else
 typedef struct {
     u16 DeviceId;
-    u32 Control_BaseAddress;
+    u64 Control_BaseAddress;
 } XByte_count_Config;
 #endif
